+++ lib/Conversion/CombToCGGI/CombToCGGI.cpp
@@ -48,7 +48,7 @@ namespace {
 Value buildSelectTruthTable(Location loc, OpBuilder &b, Value t, Value f,
                             const APInt &lut, ValueRange lutInputs) {
   int tableSize = lut.getBitWidth();
-  assert(tableSize == (1ull << lutInputs.size()));
+  assert( static_cast<unsigned long long>(tableSize) == (1ull << lutInputs.size()));
   if (tableSize == 1) {
     return lut.isZero() ? f : t;
   }
